requirements.txt
transformers
datasets
torch
pandas
transformers[torch]
accelerate
loralib
peft
trl
flash-attention
bitsandbytes
flash_attn
huggingface_hub
python-dotenv
wandb
absl-py
nltk
rouge_score